requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-Login==0.6.3
Flask-WTF==1.1.1
Flask-Migrate==4.0.5
WTForms==3.0.1
Werkzeug==2.3.7
bcrypt==4.0.1
Pillow==10.0.1
pytesseract==0.3.10
scikit-learn==1.3.0
nltk==3.8.1
textblob==0.17.1
geopy==2.3.0
pandas==2.0.3
numpy==1.24.3
python-dotenv==1.0.0
PyMySQL==1.1.0
requests==2.31.0
email-validator==2.0.0